Use a hash set for visited items instead of Zend HashTable
The main reason to switch away from Zend HashTable is that it uses PHP
runtime managed memory. The goal is to create the heap dump when we get
an OOM error (using shutdown function). If the extensions is trying to
allocate memory in PHP then the extension will trigger a second OOM
exception.

I confirmed that allocating memory directly in c does not trip a PHP
memory exception.

The hash set implementation was taken from
https://github.com/avsej/hashset.c and appears to have originated from a
couchbase project. This was done with little consideration given to the
performance or behaviours of the implementation "quick and dirty". All
unit tests still pass. It has not been tested for performance
differences.