WIP: Rework extension to traverse symbol tables zvals using stack
Made a copy of zend_ptr_stack and kept only basic functionallity and
switched allocation to use stdlib realloc() instead of the php
allocator. This is done so that memory used by the stack doesn't count
against PHP memory limit.

This is in a WIP state and needs a lot of testing and cleanup.