WIP: Troubleshooting memory leak
Found that zend_release_properties was not freeing up HashTable that is
allocated when fetching an objects properties. I haven't dug into to see
why this is, but I found that calling zend_hash_destroy() instead
ensures that the memory is freed.

After making this change, php-meminfo was able to do a full dump of the
feature cache scenario from the OOM shutdown function without triggering
a secondary OOM error. This resulted in a heap file that was 2.1GB and a
treemap html file that was 737KB with depth of 5.