Clean and changes to allow for dump during OOM error
Added a "destructive" flag which will free object properties after they
have been traversed, rather than calling zend_release_properties() which
is the recommended approach. Use the destructive flag is a bit unsafe,
so it should only be used during shutdown.

Also added an error handler to detect when an OOM error occurs. INI
settings need to be added to control if we should dump the heap during
OOM error, and where to write the output.